Add capability of modeling wind
The _drag function is duplicated for a new physics model that handles the effects simply.  Rather than drag be related to the drone velocity, it's related to the relative velocity of the drone with respect to the wind speed.
The _housekeeping function provides a random wind speed and direction for each reset of the environment.  This wind speed is between zero m/s and a maximum wind speed specified as a parameter of the environment.
The init function adds an initialization parameter for the maximum windspeed.